+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## Version 1.0.2
+
+- Fixed bad breadcrumb positioning
+- Fixed bad breadcrumbs links
+
 ## Version 1.0.1
 
 - Added option for breadcrumbs